+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - The store collection did not delete index values.
   - An insert / update / delete in a store did not globally invalidate the cache.
   - Fix collection object duplication where object was inserted with a PK = null and then updated with PK != null.
+  - Fix store's path error when displaying builder error.
 
 ## 1.6.1
 *2018-11-19*