+++ CHANGELOG.md
@@ -17,6 +17,9 @@
   - Fix collection object duplication where object was inserted with a PK = null and then updated with PK != null.
   - Fix store's path error when displaying builder error.
   - Fix store FilterUpdated event. An empty local store cannot longer send the event.
+  - Fix template compilation when imbricating rights.
+  - When deleting a value from a store, the value could have been not deleted if the same value were inserted in the meantime.
+  - Fix translations by enabling profiling.
 
 ## 1.6.1
 *2018-11-19*