+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Lunaris
 
+## 1.8.0
+- Features:
+- Fix:
+
 ## 1.7.0
 *2018-12-17*
 - Features :